perf(index): parallelize search and fix build violations
This change parallelizes similarity search and filtering in BruteForce and
LshIndex using Rayon. It also resolves build failures in WASM and MCP
feature configurations, and satisfies the 500 LOC gate for src/wasm.rs.

Key changes:
- Parallelized BruteForce and LshIndex search loops with Rayon par_iter.
- Fixed duplicate MAX_IMPORT_SIZE and LOC gate violation in src/wasm.rs.
- Resolved missing tempfile dependency for non-optional components.
- Fixed stub Persistence API mismatch in src/lib.rs for MCP feature.
- Added missing builder stubs in src/framework_builder.rs.
- Verified compilation and correctness across all targets and features.
