+++ Cargo.toml
@@ -52,7 +52,7 @@ include = [
 ]
 
 [dependencies]
-tempfile = { workspace = true, optional = true }
+tempfile = { workspace = true }
 # Serialization
 serde = { workspace = true }
 serde_json = { workspace = true }
@@ -243,7 +243,7 @@ embed-openai = ["dep:reqwest"]
 embed-voyage = ["dep:reqwest"]
 
 # ANN backends (ADR-0068)
-ann-hnsw = ["dep:hnsw_rs", "dep:tempfile"]
+ann-hnsw = ["dep:hnsw_rs"]
 ann-lsh = []
