+++ llms.txt
@@ -25,6 +25,7 @@
 - rmcp (1.7) [features: server, macros, transport-io]
 - serde
 - serde_json
+- tempfile
 - thiserror
 - tracing
 - tracing-subscriber
@@ -378,6 +379,7 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 
 ### src/framework.rs
 
+- pub const MAX_IMPORT_SIZE
 - pub struct ChaoticSemanticFramework
 - impl ChaoticSemanticFramework
 
@@ -1347,7 +1349,7 @@ homepage = "https://github.com/d-o-hub/chaotic_semantic_memory"
 
 [workspace.dependencies]
 serde = { version = "1.0.219", features = ["derive"] }
-serde_json = "1.0.149"
+serde_json = "1.0.150"
 thiserror = "2.0.11"
 tracing = "0.1.41"
 tracing-subscriber = "0.3.19"
@@ -1384,6 +1386,7 @@ include = [
 ]
 
 [dependencies]
+tempfile = { workspace = true }
 # Serialization
 serde = { workspace = true }
 serde_json = { workspace = true }
@@ -1441,11 +1444,11 @@ console_error_panic_hook = { version = "0.1", optional = false }
 
 [dev-dependencies]
 criterion = { version = "0.5.1", features = ["html_reports"] }
-tempfile = { workspace = true }
 proptest = "1.6.0"
 # CLI testing
 assert_cmd = "2.0.16"
 predicates = "3.1.3"
+mockito = "1.5.0"
 
 [[bench]]
 name = "benchmark"
