+++ src/framework.rs
@@ -21,6 +21,9 @@ use crate::singularity::{Concept, ConceptBuilder, Singularity, unix_now_secs};
 use js_sys::Date;
 
 /// Main framework for chaotic semantic memory
+/// Maximum allowed size for binary import (100MB)
+pub const MAX_IMPORT_SIZE: u64 = 100 * 1024 * 1024;
+
 pub struct ChaoticSemanticFramework {
     pub(crate) singularity: Arc<RwLock<Singularity>>,
     #[cfg(feature = "persistence")]
